+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 ### Added
 - Completion for `if`, `while`, `for`, `switch` and `try`, `catch`, `finally`.
+- Plugin's icon.
 
 ### Fixed
 - Highlight `static` keyword in non-strict mode.